requirements.txt
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.2
bcrypt==4.0.0
billiard==3.6.4.0
blinker==1.5
Brotli==1.0.9
cachelib==0.9.0
celery==5.2.7
cffi==1.15.1
chardet==5.0.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
contourpy==1.0.5
cssselect2==0.7.0
cycler==0.11.0
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.2.1
excel==1.0.0
Flask==2.1.0
Flask-Cache==0.13.1
Flask-Caching==2.0.1
Flask-Cors==3.0.10
Flask-Excel==0.0.7
Flask-Ext==0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Security-Too==5.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
fonttools==4.37.3
greenlet==1.1.3
html5lib==1.1
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
kombu==5.2.4
lml==0.1.0
MarkupSafe==2.1.1
matplotlib==3.6.0
numpy==1.23.3
packaging==21.3
passlib==1.7.4
Pillow==9.2.0
prompt-toolkit==3.0.31
pycparser==2.21
pydyf==0.3.0
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pyparsing==3.0.9
pyphen==0.13.0
python-dateutil==2.8.2
pytz==2022.2.1
redis==4.3.4
six==1.16.0
SQLAlchemy==1.4.40
texttable==1.6.4
tinycss2==1.1.1
vine==5.0.0
wcwidth==0.2.5
weasyprint==56.1
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
xlrd==2.0.1
zipp==3.8.1
zopfli==0.2.1